==> Pipelines/0.0_data_scraping_youtube_API.ipynb <==
# pip install PyTube

# pip install python-pptx

# pip install youtube-transcript-api

import glob
from Python_Functions.userFunctions import pptToText
from Python_Functions.userFunctions import youtubeTranscriptApiData

pptList = [g for g in glob.glob('../../Data/text_univ_chichago/*.pptx')]
pptDf = [pptToText(p) for p in pptList]

pptDf[0].to_csv('scraping_data/pptFlamingo.csv')
pptDf[0].head()

vidDataDfFlamingo = youtubeTranscriptApiData(searchText = 'flamingo kid education', category= 'Flamingo', i = 50)
vidDataDfFlamingo.to_csv('scraping_data/videoFlamingo.csv')
vidDataDfFlamingo.head()

pptDf[1].to_csv('scraping_data/pptPeafowl.csv')
pptDf[1].head()

vidDataDfPeafowl = youtubeTranscriptApiData(searchText = 'peafowl kid education', category= 'Peafowl', i = 50)
vidDataDfPeafowl.to_csv('scraping_data/videoPeafowl.csv')
vidDataDfPeafowl.head()

pptDf[2].to_csv('scraping_data/pptTailorBird.csv')
pptDf[2].head()

# modified
vidDataDfTailorBird = youtubeTranscriptApiData(searchText ='tailor bird for kid education',category= 'TailorBird', i = 50)
vidDataDfTailorBird.to_csv('scraping_data/videoTailorBird.csv')
vidDataDfTailorBird.head()

# vidDataDf_TailorBird = youtubeTranscriptApiData(searchText =  'tailor bird kid education', category='Tailor Bird' , i = 30)
# vidDataDf_TailorBird.to_csv('scraping_data/videoTailorBird.csv')
# vidDataDf_TailorBird.head()

pptDf[3].to_csv('scraping_data/pptAnt.csv')
pptDf[3].head()

vidDataDfAnt = youtubeTranscriptApiData(searchText ='ant kid education',category= 'Ant', i = 50)
vidDataDfAnt.to_csv('scraping_data/videoAnt.csv')
vidDataDfAnt.head()

pptDf[4].to_csv('scraping_data/pptBeetle.csv')
pptDf[4].head()

vidDataDfBeetle = youtubeTranscriptApiData(searchText = 'beetle kid education',  category= 'Beetle', i = 50)
vidDataDfBeetle.to_csv('scraping_data/videoBeetle.csv')
vidDataDfBeetle.head()
