==> Pipelines/0.5_data_engineering_cleaning_preprocessing_algorithm_based_py37.ipynb <==
import pandas as pd
import numpy as np
import glob
import os
import warnings
warnings.filterwarnings("ignore")

### Step 1 Preprocess the text Level

reference_text_df = pd.read_csv("scraping_data/Targeted_Ouput/0_raw_metadata.csv")
reference_text_df = reference_text_df.rename(columns={"raw_text": "reference_text"})

df_text = pd.read_csv("scraping_data/Targeted_Ouput/0_raw_audio_rawtxt.csv")

def topic_label_conditions(s, col = 'video_file'):
    if "ant" in s[col].lower() :
        return "insect_ant"
    elif "beetle" in s[col].lower() :
        return "insect_beetle"
    elif "bug" in s[col].lower() :
        return "insect_beetle"
    elif ("peacock" in s[col].lower()) & ("bird" in s[col].lower()) :
        return "bird_peacock"
    elif ("peafowl" in s[col].lower()) & ("bird" in s[col].lower()) :
        return "bird_peacock"
    elif ("peacock" in s[col].lower())  :
        return "bird_peacock"
    elif ("peafowl" in s[col].lower())  :
        return "bird_peacock"
    elif ("flamingo" in s[col].lower())  & ("bird" in s[col].lower()) :
        return "bird_flamingo"
    elif "flamingo" in s[col].lower() :
        return "bird_flamingo"
    elif "tailorbird" in s[col].lower() :
        return "bird_tailorbird"
    else:
        return "bird_tailorbird"

df_text['main_topics'] = df_text.apply(topic_label_conditions, axis=1)
df_text['video_file'] = df_text['video_file'].str.strip().str.replace('  ', ' ').str.replace('   ', ' ').str.replace(' ', '_').str.replace('__', '_').str.replace("&", "")


# link_df['video_name'] = link_df['video_name'].str.strip().str.replace('  ', ' ').str.replace('   ', ' ').str.replace(' ', '_').str.replace('__', '_').str.replace("&", "")


# df_text_2 = df_text.merge(link_df, how = 'left', left_on = 'video_file', right_on = 'video_name' )

reference_text_df['main_topics'] = reference_text_df.apply(lambda x: topic_label_conditions(x, col = "meta_file"), axis=1)

df_merged = df_text.merge(reference_text_df, how='left', on='main_topics' )
df_merged = df_merged.loc[:, ~df_merged.columns.isin(['video_name', 'meta_file']) ]
df_merged = df_merged.drop(['Unnamed: 0'], axis = 1)
df_merged.head()

df_merged.to_csv("scraping_data/Targeted_Ouput/text_level_df.csv", index = False)

### Step 2 Preprocess the Sentences Level

# pip install --upgrade numpy pip install --upgrade numpy

from Python_Functions.NLP_Data_Engineer_Functions import stop_words, text_cleaning, generate_words_level_timstamps,aggregate_sentences_timestamps


timestamp_texts_sentences = pd.read_csv("scraping_data/Targeted_Ouput/0_raw_audio_word_timestamp.csv")
timestamp_texts_sentences['Label'] = timestamp_texts_sentences.apply(lambda x: topic_label_conditions(x, col = "audio_source"), axis=1)

# replace all NAN value with empty space ""
timestamp_texts_sentences = timestamp_texts_sentences.fillna('')

timestamp_texts_sentences.head()

# loop this functions to each audio file timestamped extracted clips
words_timestamps_df = pd.DataFrame()
splited_clip_level = list(timestamp_texts_sentences.video_file)
for i in splited_clip_level:
    
    words_timestamps_df_temp = generate_words_level_timstamps(clip_nm  = i, timestamp_texts_sentences_sub_df = timestamp_texts_sentences )
    words_timestamps_df = pd.concat([words_timestamps_df,words_timestamps_df_temp ], axis = 0)

words_timestamps_df = words_timestamps_df.reset_index(drop = True)

# save to csv
words_timestamps_df.to_csv("scraping_data/Targeted_Ouput/1_words_timestamps_df.csv", index = None)



audio_texts = pd.read_csv("scraping_data/Targeted_Ouput/0_raw_audio_rawtxt.csv")
audio_texts.head()

# loop the function above to genrate aggreagted time range for each tokenzied sentecnes
audio_sentences_timestamp_df = pd.DataFrame()
sources = list(audio_texts.video_file)
for i in range(0,len(sources) ):
    
    audio_sentences_timestamp_temp = aggregate_sentences_timestamps(txt = audio_texts["raw_txt"][i] , words_timestamps_df_temp = words_timestamps_df[words_timestamps_df['audio_source']==sources[i]].reset_index(drop=True), audio_source = sources[i] )
    if audio_sentences_timestamp_temp.empty:
        continue
    else: 
        audio_sentences_timestamp_df = pd.concat([audio_sentences_timestamp_df, audio_sentences_timestamp_temp], axis = 0)


# sentecnes level dataset merge with the reference text
audio_sentences_timestamp_df['info_loss_perc'] = np.around(np.abs(audio_sentences_timestamp_df['diff_ratio'] - 1) * 100, 3)

audio_sentences_timestamp_df['main_topics'] = audio_sentences_timestamp_df.apply(lambda x: topic_label_conditions(x, col = "audio_source"), axis=1)
df_merged_sentences_timestamp = audio_sentences_timestamp_df.merge(reference_text_df, how='left', on='main_topics' )


df_merged_sentences_timestamp.columns

df_merged_sentences_timestamp.head()

df_merged_sentences_timestamp.to_csv("scraping_data/Targeted_Ouput/1_audio_sentences_timestamp_df.csv", index = None)



### Step 3 Spelling Correction 

# pip install https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz

# pip install spaCy==2.2

from platform import python_version

print(python_version())

# pip install --upgrade ipykernel

import en_core_web_sm
import spacy
nlp = en_core_web_sm.load()

naturallanguageprocessing = spacy.load('en')

import pandas as pd
import numpy as np
from Python_Functions.NLP_Data_Engineer_Functions import text_cleaning, stop_words
import nltk
from nltk.tokenize import sent_tokenize

import spacy
from gramformer import Gramformer
import torch
import en_core_web_sm
from nltk.tokenize import word_tokenize

nlp = en_core_web_sm.load()
nlp = en_core_web_sm.load()

nlp = spacy.load("en")
# nlp = spacy.load("en_core_web_sm")

# pip install git+https://github.com/PrithivirajDamodaran/Gramformer.git

from Python_Functions.NLP_spelling_correction import levenshtein, \
        UncommonWords, word_cnt_compare, Correct_Sentences, version_select, set_seed

sentences_extracted_pd =  pd.read_csv("scraping_data/Targeted_Ouput/1_audio_sentences_timestamp_df.csv")
sentences_extracted_pd.head()

text_extracted_pd = pd.read_csv("scraping_data/Targeted_Ouput/text_level_df.csv")
videos_files_lst = list(text_extracted_pd['video_file'])
text_extracted_pd.head()



files_identifiers = list(text_extracted_pd['video_file'].unique())

temp_txt = text_extracted_pd[text_extracted_pd['video_file'] == 'birds_intro_flamingo_v']['raw_txt'].values[0]
temp_text = temp_txt.replace(".", "")
sentences = sent_tokenize(temp_txt)
sentences_no_periods = sent_tokenize(temp_text )



# implement Gramformer API to correct text sentences 
def Correct_Sentences(file_identifier, temp_txt):
    
#     nlp = spacy.load("en_core_web_sm")
    nlp = en_core_web_sm.load()
#     set_seed(1212)
    gf = Gramformer(models = 1, use_gpu=False) # 1=corrector, 2=detector
    
    sentences = sent_tokenize(temp_txt)
    
    raw_sentences =  sentences 
    corrected_sentences = []
    Levenshtain_scores = []
    Jaccard_scores = []
    words_diff = []
    for i in range(0, len(raw_sentences)):
        raw_sentence = raw_sentences[i]
        corrected_sentence = list(gf.correct(raw_sentence, max_candidates=1))[0]
        corrected_sentences.append(corrected_sentence )
        Levenshtain_scores.append(levenshtein(seq1 = raw_sentence , seq2 = corrected_sentence))
        Jaccard_scores.append(nltk.jaccard_distance(set(raw_sentence), set(corrected_sentence) ))
        words_diff.append(word_cnt_compare(raw_sentence, corrected_sentence))
        
    # formulatae the corrected sentences back to txt chunk
    start_txt = ''
    for indx, sen in enumerate(corrected_sentences):
        if indx == 0:
            corrected_txt = start_txt + sen 
        else:
            corrected_txt = corrected_txt + " " + sen
    
    # form dataframe from all the outputs
    res_df = pd.DataFrame()
    res_df["file_name"] = [file_identifier] * len(corrected_sentences)
    res_df["raw_txt"] = [temp_txt] * len(corrected_sentences)
    res_df["raw_sentences"] = raw_sentences
    res_df["corrected_sentences"] = corrected_sentences
    res_df["corrected_txt"] = corrected_txt
    res_df["sentences_levenshtain_scores"] = Levenshtain_scores
    res_df["sentences_jaccard_scores"] = Jaccard_scores
    res_df["sentences_num_words_diff"] = words_diff
    res_df["sentences_num_words_diff_abs"] = np.abs(words_diff)
    
    raw_sentences_word_counts = [len(word_tokenize(i)) for i in list(res_df["raw_sentences"])]
    res_df["raw_sentences_word_counts"] = raw_sentences_word_counts
    res_df["correction_ratio"] = res_df["sentences_num_words_diff_abs"]/res_df["raw_sentences_word_counts"]
    

    # select to keep the corewcted txt vs keep the raw_txt
    res_df['final_corrected_version_sentences_txt'] = res_df.apply(lambda x: version_select(x.raw_sentences, x.corrected_sentences, x.correction_ratio, x.raw_sentences_word_counts ), axis=1)
    
    # formulatae the final corrected sentences back to txt chunk
    start_txt = ''
    for indx, sen in enumerate(list(res_df['final_corrected_version_sentences_txt'])):
        if indx == 0:
            final_txt = start_txt + sen 
        else:
            final_txt = corrected_txt + " " + sen
    
    res_df['final_corrected_version_sentences_txt'] = final_txt
    
    return res_df

files_identifiers = list(text_extracted_pd['video_file'].unique())

nlp = spacy.load('en')

# import spacy.cli 
# spacy.cli.download("en_core_web_md")

# pip list

result_corrected_df = pd.DataFrame()
for file in files_identifiers:
    
    temp_txt = text_extracted_pd[text_extracted_pd['video_file'] == file]['raw_txt'].values[0]
    temp_pd = Correct_Sentences(file_identifier = file, temp_txt = temp_txt)
    result_corrected_df = pd.concat([result_corrected_df, temp_pd ], axis = 0)
    
result_corrected_df.reset_index(drop=True)
