==> Pipelines/1.0_data_engineering_cleaning_preprocessing_YoutubeAPI_Merge_py37.ipynb <==
### combine, clean and format algorithm based scraped data with the API scrape data togather into 1 dataset

import pandas as pd
import numpy as np
import glob
import os

# list_of_csvs_ppt = sorted(glob.glob('scraping_data/ppt*.csv'))
list_of_csvs_API = sorted(glob.glob('scraping_data/video*.csv'))
list_of_csvs_Algo = glob.glob('scraping_data/Targeted_Ouput/*.csv')

extracted_text_df = pd.read_csv('scraping_data/Targeted_Ouput/1_transcaript_txt_sentences_level_corrected_timestamp.csv').drop(['Unnamed: 0'], axis = 1)

print(extracted_text_df.columns)
extracted_text_df.head()


refernece_txt_df = extracted_text_df[['main_topics', 'reference_text']].drop_duplicates()

sorted(list(extracted_text_df['main_topics'].unique()))

['insect_ant', 'insect_beetle','bird_flamingo', 'bird_peacock', 'bird_tailorbird' ]

## 1. read each API scraped video
## 2. calculate the end time stamp in each sentences
## 3. label each sentences with the main topic
## 4. combin each sentences togather into a whole text transcript with "." as sperators
## 5. create a new column called "transcript_text" to contain the whole transcriop text in each senetcnal level rows
## 6. finally put each df everything toagther into whole big DF

API_main_df = pd.DataFrame()

main_topics_lst = ['insect_ant', 'insect_beetle','bird_flamingo', 'bird_peacock', 'bird_tailorbird' ]

for idx in range(0, len(list_of_csvs_API)):
    
    api_temp = pd.read_csv(list_of_csvs_API[idx])
    
    # calculate the end timestamp
    api_temp['end'] = api_temp['start'] + api_temp['duration']
    
    # label each sentences with the main topic
    api_temp['main_topics'] = main_topics_lst[idx]
    
    # combine each sentences togather into a whole text transcript with "." as sperators
    video_ids = list(api_temp['Video ID'].unique())
    trans_temp_df_dict = {"Video_ID" : [], "Transcript_Txt": []}
    video_trans_df = pd.DataFrame()
        
    for video in video_ids:
        video_sentences = list(api_temp[api_temp['Video ID'] == video]['text'])
        transcript = ""

        for sentence in video_sentences:
            
            if video_sentences.index(sentence) == 0:
                transcript = transcript + sentence
            else:
                transcript = transcript + ". " + sentence
        trans_temp_df_dict["Video_ID"].append(video)
        trans_temp_df_dict["Transcript_Txt"].append(transcript)
        
    
    video_trans_df = pd.DataFrame.from_dict(trans_temp_df_dict)
    
    # 5. create a new column called "transcript_text" to contain the whole transcriop text in each senetcnal level rows
    api_temp = pd.merge(api_temp, video_trans_df, 'left', left_on = "Video ID" , right_on = "Video_ID")
    
    # finally put each df everything toagther into whole big DF
    API_main_df = pd.concat([API_main_df, api_temp], axis = 0).drop(['Unnamed: 0','Video_ID' ], axis = 1)
    

# add refenrece text
API_main_df = pd.merge(API_main_df,refernece_txt_df, 'left', left_on = 'main_topics' , right_on = 'main_topics')

# rename and reorder some of the API_name dataframe

API_main_df_col = ['text', 'start', 'duration', 'Video ID', 'Title', 'Age Restricted',
       'Length (min)', 'Views (thous)', 'Category', 'end', 'main_topics',
       'Transcript_Txt', 'reference_text']

API_main_df_new = ['final_corrected_version_sentences_txt', 'sentence_level_timstamp_min_sec', 'duration', 'Video_ID', 'Title', 'Age_Restricted',
       'Length_(min)', 'Views_(thous)', 'Category', 'sentence_level_timstamp_max_sec', 'main_topics',
       'final_corrected_version_txt', 'reference_text']

change_col_name_map = dict(zip(API_main_df_col, API_main_df_new))
API_main_df = API_main_df.rename(columns=change_col_name_map)

API_main_df['sentence_level_timstamp_min_minute'] = np.round(API_main_df['sentence_level_timstamp_min_sec']/60, 6)

API_main_df['sentence_level_timstamp_max_minute'] = np.round(API_main_df['sentence_level_timstamp_max_sec']/60, 6)

API_main_df = API_main_df[['Title', 'Video_ID', 'Category', 'Age_Restricted', \
             'final_corrected_version_sentences_txt', 'final_corrected_version_txt',\
             'sentence_level_timstamp_min_sec', 'sentence_level_timstamp_max_sec',\
             'sentence_level_timstamp_min_minute', 'sentence_level_timstamp_max_minute', \
             'duration', 'Length_(min)', 'Views_(thous)', \
            'main_topics', 'reference_text']]


print(API_main_df.columns)

API_main_df.head()

API_main_df.columns

# rename some of the columns in extracted_text_df
extracted_text_df_processed = extracted_text_df.copy()
extracted_text_df_processed['Video_ID'] = extracted_text_df_processed['file_name']
extracted_text_df_processed['Views_(thous)'] = 0
extracted_text_df_processed['Category'] = ''
extracted_text_df_processed['duration'] = 0
extracted_text_df_processed['Length_(min)'] = 0
extracted_text_df_processed['Age_Restricted'] = False
extracted_text_df_processed['Title'] = extracted_text_df_processed['file_name']

extracted_text_df_processed = extracted_text_df_processed[[
    'Title', 'Video_ID', 'Category', 'Age_Restricted',\
    'final_corrected_version_sentences_txt', 'final_corrected_version_txt',\
       'sentence_level_timstamp_min_sec','sentence_level_timstamp_max_sec', \
    'sentence_level_timstamp_min_minute','sentence_level_timstamp_max_minute',\
    'duration', 'Length_(min)', 'Views_(thous)', 'main_topics','reference_text'
    
]]

extracted_text_df_processed.columns
extracted_text_df_processed.head()

## combine and concat the API downlaod text with the algo scraped text into a final DF to be later feed into the models
main_df = pd.concat([API_main_df, extracted_text_df_processed ], axis = 0)
main_df.head()

### clean the reference txt and mark the manually extracted subtopics

from Python_Functions.NLP_Data_Engineer_Functions import stopwords_generator, lemmatize_stemming, TF_whole_text, TF_IDF_text_doc, topic_extract, \
                          lemmatize_preprocess, text_cleaning_sub, text_cleaning, stop_words


## from the PPT

birds_tailorbird_manual_keywords = ["tailorbird", 'colour', 'color' "appearances",\
                                    "beak", "food", 'voice', 'habitat', 'nest', "egg"  ]

birds_tailorbird_manual_useless_sentences = ['Hey I am Tinky  a Tailorbird. Do you know that I sing melodious songs? Let’s learn more about me!'] 


birds_peacock_manual_keywords = ["peacock", "peafowl", 'feature', 'specialty', 'train', 'crest', 'peahen',
                                 'green peafowl', 'white peafowl', 'colour', 'color' "appearances",\
                                    "beak", "food", 'voice', 'habitat', 'nest', "egg"  ]

birds_peacock_manual_useless_sentences = ['Source: peacocksuk.com', 'Source: pinterest.com', 'Source: natureconservation.in', \
                                         'Source: en.wikipedia.org'] 


birds_flamingo_manual_keywords = ["flamingo", 'feature', 'pink', 'specialty', 'train', 'carotenoids'
                                 'lesser flamingo', 'american flamingo',  'andean flamingo',\
                                 'chilean flamingo' , 'james flamingo', 'colour', 'color' "appearances",\
                                    "beak", "food", 'voice', 'habitat', 'nest', "egg", 'milk'  ]

birds_flamingo_manual_useless_sentences = ['Source: ebird.org', 'Source: en.wikipedia.org, commons.wikimedia.org'] 


insect_ant_manual_keywords = ["ant", 'queen', 'male', "drone", 'worker', 'social', 'princess', 'soldier', 'larva', 'larvae',\
                              'castes', 'black ants', 'appearance', 'habit', 'carpenter', 'weaver', 'leafcutter'\
                              "Odorous house", "food", 'nest', "egg", 'Life cycle'  ]

insect_ant_manual_useless_sentences = ['GLOSSARY: 1.Honeydew: a sweet sticky substance produced by insects while they feed on the plant sap. 2.Plant secretion: the produce of plants such as gums resins etc.'] 


insect_beetle_manual_keywords = ["beetle", 'elytra', 'wing', "ladybug", 'appearance', 'habit', 'scarab', 'mango stem borer', 'weevil', 'dung',\
                              'castes', 'black ants', 'appearance', 'habit', 'carpenter', 'weaver', 'leafcutter'\
                              "Odorous house", "food", 'nest', "egg", 'Life cycle'  ]

insect_beetle_manual_useless_sentences = ['We see many insects around us. Some have wings like mosquitoes, flies, butterflies, etc., while others do not have wings like ants, caterpillars, etc.',\
                                         '\\t', 'GLOSSARY: Species: A group of living organisms that are alike in their features and behaviour. Snout: a long nose. Millet: A type of cereal.' ] 

strings_to_remove = birds_tailorbird_manual_useless_sentences + birds_peacock_manual_useless_sentences + birds_flamingo_manual_useless_sentences +\
             insect_ant_manual_useless_sentences + insect_beetle_manual_useless_sentences

manual_reference_topics = {'bird_tailorbird' :  birds_tailorbird_manual_keywords,
                           'bird_peacock' : birds_peacock_manual_keywords, 
                           'bird_flamingo' : birds_flamingo_manual_keywords,
                          'insect_ant' : insect_ant_manual_keywords, 
                           'insect_beetle' : insect_beetle_manual_keywords}


import re

def remove_urls (vTEXT):
    vTEXT = re.sub(r'(https?:\/\/)(\s)*(www\.)?(\s)*((\w|\s)+\.)*([\w\-\s]+\/)*([\w\-]+)((\?)?[\w\s]*=\s*[\w\%&]*)*', '', vTEXT, flags=re.MULTILINE)
    vTEXT = re.sub(r'(https|http)?:\/\/(\w|\.|\/|\?|\=|\&|\%)*\b', '', vTEXT, flags=re.MULTILINE)
    vTEXT = vTEXT.replace("'/'","").replace("/","").replace("'https://'", '').replace("'[Music]'", '')
    return(vTEXT)

def form_txt_chunks(txt_lst, sep = ' '):
    
    if txt_lst != []:
        start_txt = ''
        for indx, sen in enumerate(txt_lst):
            if indx == 0:
                put_togather = start_txt + sen 
            else:
                put_togather = put_togather + sep + sen
        return put_togather
    else: return ""
        
def clean_txt_chunk(txt):
    
    sep_lst = txt.split(",")
    cleaned_lst = [i.replace("'", '').strip().replace("[", '').replace("]", '').replace(".jpg", '').replace(".html", '') for i in sep_lst]
    put_togather = form_txt_chunks(txt_lst = cleaned_lst, sep = ' ')

    return remove_urls (vTEXT = put_togather)


refer_txt_df = main_df[['reference_text', 'main_topics']].drop_duplicates().reset_index(drop = True)

refer_txt_df["reference_text_cleaned"] = refer_txt_df['reference_text'].apply(clean_txt_chunk)

# remove the useless sentecnes that could be affect the learning of the concepts/facts
refer_txt_df['reference_text_cleaned_2x'] = refer_txt_df['reference_text_cleaned']
for strg in strings_to_remove:
    refer_txt_df['reference_text_cleaned_2x'] = refer_txt_df['reference_text_cleaned_2x'].str.replace(strg, "")

# apply the manally selected sub-topic key words to the data frame
main_topics_lst = []
manual_topics_string = []
for key, manual_topics in manual_reference_topics.items():
    manual_topics_string.append(form_txt_chunks(txt_lst = manual_topics, sep = ','))
    main_topics_lst.append(key)


df_manual_topics = pd.DataFrame(list(dict(zip(main_topics_lst, manual_topics_string)).items()),columns = ['main_topics','manual_subtopics']) 
refer_txt_df = pd.merge(refer_txt_df, df_manual_topics, "left", 'main_topics')

refer_txt_df_new = refer_txt_df[[ 'reference_text_cleaned_2x', 'manual_subtopics', 'main_topics']]


# form the sentences level from the reference text
refertxt_sentences_lvl_df = pd.DataFrame()
for index, row in refer_txt_df_new.iterrows():
    refertxt_sentences_lvl_df_temp = pd.DataFrame()
    txt = row['reference_text_cleaned_2x']
    clss = row['main_topics']
    manual_subtopic = row['manual_subtopics']
    
    cleaned_temp = text_cleaning(txt = txt, stop_words = stop_words)

    cleaned_words_agg  = form_txt_chunks(txt_lst = cleaned_temp[1], sep = ',')
    vocabulary_agg  = form_txt_chunks(txt_lst =  cleaned_temp[2], sep = ',') # unique words
    tokens_sentences_lst = cleaned_temp[3]
    
    refertxt_sentences_lvl_df_temp['sentences'] = tokens_sentences_lst
    refertxt_sentences_lvl_df_temp['unique_vocabulary_agg'] = vocabulary_agg
    refertxt_sentences_lvl_df_temp['cleaned_words_agg'] = cleaned_words_agg
    refertxt_sentences_lvl_df_temp['reference_text_cleaned']  = txt
    refertxt_sentences_lvl_df_temp['manual_subtopics']  = manual_subtopic
    refertxt_sentences_lvl_df_temp['main_topics'] = clss
    
    refertxt_sentences_lvl_df = pd.concat([refertxt_sentences_lvl_df, refertxt_sentences_lvl_df_temp], axis = 0)
    
    

main_df = pd.merge(main_df, refertxt_sentences_lvl_df[['reference_text_cleaned', 'main_topics' ]].drop_duplicates(),\
                   'left', 'main_topics')



main_df.shape

refertxt_sentences_lvl_df.to_csv("ready_for_model_dataset/refertxt_sentences_lvl_df.csv")

main_df.to_csv("ready_for_model_dataset/ready_for_model.csv")
