==> Pipelines/2.0_coverage_model_py38.ipynb <==
## measure coverage in text and sentences level against the reference text

import pandas as pd
#!python -m spacy download en_core_web_lg
from Python_Functions.userFunctions import countTfidfWord2VecScore_between_2_text
# import en_core_web_sm
# nlp=en_core_web_sm.load()

import spacy
import en_core_web_lg
nlp = en_core_web_lg.load()
import pandas as pd
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import cosine_similarity
import numpy as np


# ready to model csv
Df = pd.read_csv('ready_for_model_dataset/ready_for_model.csv').drop(['Unnamed: 0'], axis = 1)
print(Df.shape)
print(Df.columns)
Df.head()

#### measure cosline simialrity between 2 columns

Transcript File level vs Sentences Level

- compare transcript text vs reference text
- compare transcript sentences vs reference text


- compare transcript text vs unique_vocabulary_agg(refernece text)
- compare transcript sentences vs unique_vocabulary_agg(refernece text)


- compare transcript text vs cleaned_words_agg(refernece text)
- compare transcript sentences vs cleaned_words_agg(refernece text)


Df_txt_lvl = Df[['Title', 'Video_ID', 'main_topics',  'final_corrected_version_txt', 'reference_text_cleaned']].drop_duplicates()
print(Df_txt_lvl.shape)
vid_ids = list(Df_txt_lvl.Video_ID.unique())
Df_sentences_lvl = Df[['Title', 'Video_ID', 'main_topics', 'final_corrected_version_sentences_txt', 'reference_text_cleaned']].drop_duplicates()
print(Df_sentences_lvl.shape)
Df_sentences_unique = Df_sentences_lvl[['Video_ID', 'final_corrected_version_sentences_txt']].drop_duplicates()
sentences = list(Df_sentences_unique.final_corrected_version_sentences_txt)
vid_ids_sent = list(Df_sentences_unique.Video_ID)
vid_sentence_combo = list(zip(vid_ids_sent,sentences))

#### Based on different video id, compute the coverage score between transcript vs reference text

def apply_cosine_similarity(Df, new_doc_col, reference_doc_col, vid_ids = None, sentence_lvl = False, vid_sentence_combo = None):
    if sentence_lvl == False:
        result_df = pd.DataFrame()
        for Id in vid_ids:
            row_temp_unique_df = Df[Df['Video_ID'] == Id][['Video_ID', new_doc_col, reference_doc_col, 'main_topics']].drop_duplicates()
            new_doc_txt = list(row_temp_unique_df[new_doc_col])[0]
            reference_doc_txt = list(row_temp_unique_df[reference_doc_col])[0]
            try:
                coverage_scores = countTfidfWord2VecScore_between_2_text(new_doc_txt = new_doc_txt, reference_doc_txt = reference_doc_txt )
            except:
                coverage_scores = [0.0, 0.0, 0.0]
            row_temp_unique_df['Count_Vector_Cosine_Similarity'] = coverage_scores[0]
            row_temp_unique_df['TFIDF_Vector_Cosine_Similarity'] = coverage_scores[1]
            row_temp_unique_df['Word_Vector_Cosine_Similarity'] = coverage_scores[2]
            row_temp_unique_df['Avg_Cosine_Similarity'] = row_temp_unique_df[[
                'Count_Vector_Cosine_Similarity',
                'TFIDF_Vector_Cosine_Similarity',
                'Word_Vector_Cosine_Similarity']].mean(axis=1)
            result_df = pd.concat([result_df,row_temp_unique_df], axis = 0).sort_values(by = 'Avg_Cosine_Similarity', ascending = False)
        return result_df
    else:
        result_df = pd.DataFrame()
        for combo in vid_sentence_combo:
            row_temp_unique_df = Df[(Df['Video_ID'] == combo[0]) & (Df[new_doc_col] == combo[1])][['Video_ID', new_doc_col, reference_doc_col, 'main_topics']].drop_duplicates()
            new_doc_txt = list(row_temp_unique_df[new_doc_col])[0]
            reference_doc_txt = list(row_temp_unique_df[reference_doc_col])[0]
            try:
                coverage_scores = countTfidfWord2VecScore_between_2_text(new_doc_txt = new_doc_txt, reference_doc_txt = reference_doc_txt )
            except:
                coverage_scores = [0.0, 0.0, 0.0]
            row_temp_unique_df['Count_Vector_Cosine_Similarity'] = coverage_scores[0]
            row_temp_unique_df['TFIDF_Vector_Cosine_Similarity'] = coverage_scores[1]
            row_temp_unique_df['Word_Vector_Cosine_Similarity'] = coverage_scores[2]
            row_temp_unique_df['Avg_Cosine_Similarity'] = row_temp_unique_df[[
                'Count_Vector_Cosine_Similarity',
                'TFIDF_Vector_Cosine_Similarity',
                'Word_Vector_Cosine_Similarity']].mean(axis=1)
            result_df = pd.concat([result_df,row_temp_unique_df], axis = 0).sort_values(by = 'Avg_Cosine_Similarity', ascending = False)
        return result_df    

Df_txt_lvl_txt_vs_reference_text = apply_cosine_similarity(
    Df = Df_txt_lvl,
    vid_ids = vid_ids,
    new_doc_col = 'final_corrected_version_txt',
    reference_doc_col = 'reference_text_cleaned'
)

Df_txt_lvl_txt_vs_reference_text.head()

##### Insight: given 5 main topics, what are their top 5 videos transcript text based on the coverage score?

Df_txt_lvl_txt_vs_reference_text.to_csv('model_output_dataset/coverage_measure_output_text.csv')

main_topics_Lst = list(Df_txt_lvl_txt_vs_reference_text.main_topics.unique())

for topic in main_topics_Lst:
    print()
    print('top 10 video under main topic: {} :'.format(topic))
    temp_df = (Df_txt_lvl_txt_vs_reference_text[Df_txt_lvl_txt_vs_reference_text['main_topics'] == topic]
     .sort_values(by = 'Avg_Cosine_Similarity', ascending = False)
    )
    print()
    print(temp_df.head(10)['Video_ID'])

#### Based on different video id, compute the coverage score between sentence vs reference text

# new_doc_col = 'final_corrected_version_sentences_txt'
# reference_doc_col = 'reference_text_cleaned'


# result_df = pd.DataFrame()
# for combo in vid_sentence_combo:
# #     print(combo[0])
# #     print(combo[1])
#     row_temp_unique_df = Df_sentences_lvl[(Df_sentences_lvl['Video_ID'] == combo[0]) & \
#                                           (Df_sentences_lvl[new_doc_col] == combo[1])][\
#                                         ['Video_ID', new_doc_col, reference_doc_col, 'main_topics']].drop_duplicates()
# #     print(row_temp_unique_df)
#     new_doc_txt = list(row_temp_unique_df[new_doc_col])[0]
# #     print(new_doc_txt)
#     reference_doc_txt = list(row_temp_unique_df[reference_doc_col])[0]
#     try:
#         coverage_scores = countTfidfWord2VecScore_between_2_text(new_doc_txt = new_doc_txt, reference_doc_txt = reference_doc_txt )
#     except:
#         coverage_scores = [0.0, 0.0, 0.0]
#     print(coverage_scores)
#     row_temp_unique_df['Count_Vector_Cosine_Similarity'] = coverage_scores[0]
#     row_temp_unique_df['TFIDF_Vector_Cosine_Similarity'] = coverage_scores[1]
#     row_temp_unique_df['Word_Vector_Cosine_Similarity'] = coverage_scores[2]
#     row_temp_unique_df['Avg_Cosine_Similarity'] = row_temp_unique_df[[
#         'Count_Vector_Cosine_Similarity',
#         'TFIDF_Vector_Cosine_Similarity',
#         'Word_Vector_Cosine_Similarity']].mean(axis=1)
#     result_df = pd.concat([result_df,row_temp_unique_df], axis = 0).sort_values(by = 'Avg_Cosine_Similarity', ascending = False)
# result_df    



Df_sentences_lvl_txt_vs_reference_text = apply_cosine_similarity(
    Df = Df_sentences_lvl,
    sentence_lvl = True,
    vid_sentence_combo = vid_sentence_combo,
    new_doc_col = 'final_corrected_version_sentences_txt',
    reference_doc_col = 'reference_text_cleaned'
)

Df_sentences_lvl_txt_vs_reference_text.head()

##### Insights: Given a video, what are the top 5 sentences under that video ?

np.logspace(0,-9, num=50)

Df_sentences_lvl_txt_vs_reference_text.to_csv('model_output_dataset/coverage_measure_output_sentences.csv')

(Df_sentences_lvl_txt_vs_reference_text[Df_sentences_lvl_txt_vs_reference_text['Video_ID'] == 'FMjppU6AKBQ']
 .sort_values(by = 'Avg_Cosine_Similarity', ascending = False)
 .head()['final_corrected_version_sentences_txt']
)
