==> Pipelines/2.5_compare_vid_vs_ppt_csv_file_coverage_model_Deprioirtized_py38.ipynb <==
import pandas as pd
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)
pd.set_option('max_colwidth', None)
from nltk.corpus import stopwords
from string import punctuation
from sklearn.metrics.pairwise import cosine_similarity
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer

def remove_punctuation(dirty_string):
    non_words = list(punctuation)
    for word in non_words:
        dirty_string = dirty_string.replace(word, '')
    return dirty_string

def remove_stop_words(dirty_text):
    non_words = list(punctuation)
    language_stopwords = stopwords.words('english')
    cleaned_text = ''
    for word in dirty_text.split():
        if word in language_stopwords or word in non_words:
            continue
        else:
            cleaned_text += word + ' '
    return cleaned_text

def process_file(file_content):
    file_content = file_content.lower()
    file_content = remove_punctuation(file_content)
    file_content = remove_stop_words(file_content)
    return file_content

def countTfidfScore(ppt,vid):
    ppt = pd.read_csv(ppt)
    pptClean = process_file(str(ppt['PPT Sub Text']))
    vid = pd.read_csv(vid)
    vidGroup = [vid[(vid['Video ID'] == x)]['text'] for x in vid['Video ID'].unique()]

    vidScore = []
    for v in vidGroup:
        vidClean = process_file(str(v))
        vectorizer = CountVectorizer()
        X = vectorizer.fit_transform([pptClean,vidClean])
        vidScore.append(cosine_similarity(X,X)[0,1])
    countScoreDf = pd.concat([pd.DataFrame(vid['Video ID'].unique(), columns=['Video ID']), pd.DataFrame(vidScore, columns=['Count Vector Cosine Similarity'])], axis=1)
    vidScore = []
    for v in vidGroup:
        vidClean = process_file(str(v))
        vectorizer = TfidfVectorizer()
        X = vectorizer.fit_transform([pptClean,vidClean])
        vidScore.append(cosine_similarity(X,X)[0,1])
    countTfidfScoreDf = pd.concat([countScoreDf, pd.DataFrame(vidScore, columns=['TFIDF Vector Cosine Similarity'])], axis=1)
    return countTfidfScoreDf

countTfidfScore('scraping_data/pptFlamingo.csv','scraping_data/videoFlamingo.csv')

countTfidfScore('scraping_data/pptPeafowl.csv','scraping_data/videoPeafowl.csv')

countTfidfScore('scraping_data/pptTailorBird.csv','scraping_data/videoTailorBird.csv')

countTfidfScore('scraping_data/pptAnt.csv','scraping_data/videoAnt.csv')

countTfidfScore('scraping_data/pptBeetle.csv','scraping_data/videoBeetle.csv')
