==> Pipelines/2.75_relavance_model_Implementation_py38.ipynb <==
import sklearn
import pandas as pd
import numpy as np
import nltk
import wordcloud
import matplotlib.pyplot as plt

from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import CountVectorizer, HashingVectorizer, TfidfTransformer, TfidfVectorizer
from sklearn.naive_bayes import MultinomialNB
from sklearn.metrics import classification_report, confusion_matrix, accuracy_score
from sklearn.linear_model import LogisticRegression, SGDClassifier
from sklearn.model_selection import cross_validate
from sklearn import metrics
from sklearn.ensemble import  VotingClassifier
import pickle
from Python_Functions.NLP_Model_Functions import stopwords_generator, lemmatize_stemming, TF_whole_text, TF_IDF_text_doc, topic_extract, \
                          lemmatize_preprocess, text_cleaning_sub, text_cleaning, stop_words

## Apply Supervised Learning on the Topics Modeling - Main Topic Classification

1. Word freq under a certain topics: calculate the ratio of a words appears among all the words in a documents of topics

2. Topic Classification with Niave bayes


refer_model = pickle.load(open("models/trained_models/NaiveBeyasClassifier_ReferenceText.pckl","rb"))

classfier_mdl = refer_model[0]
vectorizer_train = refer_model[1]

df = pd.read_csv('ready_for_model_dataset/df_test.csv').drop(['Unnamed: 0'], axis = 1)
df.head()


# encode the classes to predicted
df['main_topics_coded'] = df.main_topics.map({'bird_tailorbird' : 0, 'insect_ant': 1, 'bird_peacock' :2, 'insect_beetle' : 3,
       'bird_flamingo': 4})

Transcript_text = df[['final_corrected_version_txt', 'main_topics_coded', 'coverage_quality']].drop_duplicates().reset_index(drop = True)

Transcript_text.head()

X_video_transc_txt = Transcript_text['final_corrected_version_txt'].values
y_video_transc_txt = Transcript_text['main_topics_coded'].values
X_sentences_video_transc_txt = df['final_corrected_version_sentences_txt'].values
y_sentences_video_transc_txt = df['main_topics_coded'].values

X_test_dtm_txt = vectorizer_train.transform(X_video_transc_txt)
X_test_dtm_txt

X_test_dtm_sentence = vectorizer_train.transform(X_sentences_video_transc_txt)
X_test_dtm_sentence

classfier_mdl.predict(X_test_dtm_txt)

y_pred_txt = classfier_mdl.predict(X_test_dtm_txt)
y_pred_proba_txt = classfier_mdl.predict_proba(X_test_dtm_txt)

y_pred_sentence =  classfier_mdl.predict(X_test_dtm_sentence)
y_pred_proba_sentence =  classfier_mdl.predict_proba(X_test_dtm_sentence)

class_proba_df_txt = pd.DataFrame(y_pred_proba_txt, columns = ['txt_0_proba','txt_1_proba','txt_2_proba','txt_3_proba','txt_4_proba'])
Transcript_text = pd.concat([Transcript_text, class_proba_df_txt], axis = 1)
Transcript_text['classified_topic_coded_txt'] = y_pred_txt
Transcript_text.head(10)

class_proba_df_sentence = pd.DataFrame(y_pred_proba_sentence, columns = ['sentence_0_proba','sentence_1_proba','sentence_2_proba','sentence_3_proba','sentence_4_proba'])
df = pd.concat([df, class_proba_df_sentence], axis = 1)
df['classified_topic_coded_sentences'] = y_pred_sentence
df.head(5)

print(classification_report(y_video_transc_txt, y_pred_txt))

print(classification_report(y_pred_sentence, y_sentences_video_transc_txt))

# merge the txt level nd sentences level

df_rank = pd.merge(df, Transcript_text, 'left', 'final_corrected_version_txt')
df_rank.head()

df_rank.columns

### Implement the extraction of the Insights from the classifier model 

- first pick out the correctly classified rows and label it with a new columns "Correctly classified" (done)
- correctly classified text/sentences will be helped ranked later from the topic extraction algorothm (HOPFULLY) (pending ......)
- compare the highsest probability of the classified class vs the actual class probaility 
  * the difference between them, the samller the better
  * the smaller or appraoch to 0 ==> better text quality under a specific class
  * the bigger the probability difference ==> the worse on transcript quality under a specific class



#### Lesson: what is important to have a better trained classifed model from the refernece text

- do we use perfect text to train the classiifer model? or we can add more variance of text to make the model more generlelziable? 

- the purposes of the classifier model is to give an estimate on if the transcript text belongs to a class?

- assumption: the classifier label from the transcript are assumed in the first place, our purpose is to figure out a way to help us measure the legitimacy of the main class from each video transcript 


df_rank['if_classified_correctly_txt'] = df_rank['classified_topic_coded_txt'] == df_rank['main_topics_coded_x']
df_rank['if_classified_correctly_sentences'] = df_rank['classified_topic_coded_sentences'] == df_rank['main_topics_coded_x']

df_rank['max_proba_sentence'] = df_rank[['sentence_0_proba', 'sentence_1_proba', 'sentence_2_proba','sentence_3_proba', 'sentence_4_proba']].max(axis=1)

df_rank['max_proba_txt'] = df_rank[['txt_0_proba', 'txt_1_proba', 'txt_2_proba', 'txt_3_proba','txt_4_proba']].max(axis=1)


gaps_sentences = []
sentence_actual_main_topic_classified_proba = []
proba_col_lst = ['sentence_0_proba', 'sentence_1_proba', 'sentence_2_proba','sentence_3_proba', 'sentence_4_proba']

for index, row in df_rank.iterrows():
    
    if row['if_classified_correctly_sentences'] == False:
        actual_lbl = str(row['main_topics_coded_x'])
        actual_target_col = [i for i in proba_col_lst if actual_lbl in i][0]
        
        gaps_sentences.append(np.round(1 - row[actual_target_col] , 6))
        sentence_actual_main_topic_classified_proba.append(row[actual_target_col])

    else:
        gaps_sentences.append(np.round(1 - row['max_proba_sentence'], 6)) 
        sentence_actual_main_topic_classified_proba.append(row['max_proba_sentence'])
        
gaps_txt = []
txt_actual_main_topic_classified_proba = []
proba_col_lst = ['txt_0_proba', 'txt_1_proba', 'txt_2_proba', 'txt_3_proba','txt_4_proba']

for index, row in df_rank.iterrows():
    
    if row['if_classified_correctly_txt'] == False:
        actual_lbl = str(row['main_topics_coded_x'])
        actual_target_col = [i for i in proba_col_lst if actual_lbl in i][0]
        
#         gaps_txt.append(np.round(np.abs(row['max_proba_txt'] - row[actual_target_col] ), 6))
        gaps_txt.append(np.round(np.abs(1 - row[actual_target_col] ), 6))
        txt_actual_main_topic_classified_proba.append(row[actual_target_col])

    else:
        gaps_txt.append(np.round(1 - row['max_proba_txt'], 6)) 
        txt_actual_main_topic_classified_proba.append(row['max_proba_txt'])
        
df_rank['sentence_classified_vs_actual_gap'] = gaps_sentences
df_rank['txt_classified_vs_actual_gap'] = gaps_txt

df_rank['sentence_main_topic_classified_proba'] = sentence_actual_main_topic_classified_proba
df_rank['txt_main_topic_classified_proba'] = txt_actual_main_topic_classified_proba
        

### Implement the extraction of the Insights from the classifier model 

- first pick out the correctly classified rows and label it with a new columns "Correctly classified" (done)
- correctly classified text/sentences will be helped ranked later from the topic extraction algorothm (HOPFULLY) (pending ......)
- compare the highsest probability of the classified class vs the actual class probaility 
  * the difference between them, the samller the better
  * the smaller or appraoch to 0 ==> better text quality under a specific class
  * the bigger the probability difference ==> the worse on transcript quality under a specific class



#### Lesson: what is important to have a better trained classiifed model from the refernece text

- do we use perfect text to train the classiifer model? or we can add more variance of text to make the model more generlelziable? 

- the purposes of the classifier model is to give an estimate on if the transcript text belongs to a class?

- assumption: the classifier label from the transcript are assumed in the first place, our purpose is to figure out a way to help us measure the legitimacy of the main class from each video transcript 


df_rank['if_classified_correctly_txt'] = df_rank['classified_topic_coded_txt'] == df_rank['main_topics_coded_x']
df_rank['if_classified_correctly_sentences'] = df_rank['classified_topic_coded_sentences'] == df_rank['main_topics_coded_x']

df_rank['max_proba_sentence'] = df_rank[['sentence_0_proba', 'sentence_1_proba', 'sentence_2_proba','sentence_3_proba', 'sentence_4_proba']].max(axis=1)

df_rank['max_proba_txt'] = df_rank[['txt_0_proba', 'txt_1_proba', 'txt_2_proba', 'txt_3_proba','txt_4_proba']].max(axis=1)



gaps_sentences = []
sentence_actual_main_topic_classified_proba = []
proba_col_lst = ['sentence_0_proba', 'sentence_1_proba', 'sentence_2_proba','sentence_3_proba', 'sentence_4_proba']

for index, row in df_rank.iterrows():
    
    if row['if_classified_correctly_sentences'] == False:
        actual_lbl = str(row['main_topics_coded_x'])
        actual_target_col = [i for i in proba_col_lst if actual_lbl in i][0]
        
        gaps_sentences.append(np.round(1 - row[actual_target_col] , 6))
        sentence_actual_main_topic_classified_proba.append(row[actual_target_col])

    else:
        gaps_sentences.append(np.round(1 - row['max_proba_sentence'], 6)) 
        sentence_actual_main_topic_classified_proba.append(row['max_proba_sentence'])
        
gaps_txt = []
txt_actual_main_topic_classified_proba = []
proba_col_lst = ['txt_0_proba', 'txt_1_proba', 'txt_2_proba', 'txt_3_proba','txt_4_proba']

for index, row in df_rank.iterrows():
    
    if row['if_classified_correctly_txt'] == False:
        actual_lbl = str(row['main_topics_coded_x'])
        actual_target_col = [i for i in proba_col_lst if actual_lbl in i][0]
        
#         gaps_txt.append(np.round(np.abs(row['max_proba_txt'] - row[actual_target_col] ), 6))
        gaps_txt.append(np.round(np.abs(1 - row[actual_target_col] ), 6))
        txt_actual_main_topic_classified_proba.append(row[actual_target_col])

    else:
        gaps_txt.append(np.round(1 - row['max_proba_txt'], 6)) 
        txt_actual_main_topic_classified_proba.append(row['max_proba_txt'])
        
df_rank['sentence_classified_vs_actual_gap'] = gaps_sentences
df_rank['txt_classified_vs_actual_gap'] = gaps_txt

df_rank['sentence_main_topic_classified_proba'] = sentence_actual_main_topic_classified_proba
df_rank['txt_main_topic_classified_proba'] = txt_actual_main_topic_classified_proba
        

df_rank.head()

print(df_rank.drop_duplicates().shape)
print(df_rank.shape)

df_rank.to_csv("model_output_dataset/video_tarnscript_main_topic_classifier_rank.csv")

## Apply Unsupervised Learning on the topics - Sub-Topics Extraction

Apply the LDA model to the new text/sentences documents from video transcript to get the topic probability distribution for a new document based on a trained model


https://radimrehurek.com/gensim/models/ldamodel.html

https://towardsdatascience.com/topic-modeling-and-latent-dirichlet-allocation-in-python-9bf156893c24

from Python_Functions.NLP_Model_Functions import stopwords_generator, lemmatize_stemming, TF_whole_text, TF_IDF_text_doc, topic_extract, \
                          lemmatize_preprocess, text_cleaning_sub, text_cleaning, stop_words

import gensim
from gensim.utils import simple_preprocess
from gensim.parsing.preprocessing import STOPWORDS
from nltk.stem import WordNetLemmatizer, SnowballStemmer
from nltk.stem.porter import *
import numpy as np
# np.random.seed(2018)

import nltk
nltk.download('wordnet')

def lemmatize_stemming(text):
    return SnowballStemmer.stem(WordNetLemmatizer().lemmatize(text, pos='v'))
def preprocess(text):
    result = []
#     print(gensim.utils.simple_preprocess(text))
    for token in gensim.utils.simple_preprocess(text):
#         print(token)
        if token not in gensim.parsing.preprocessing.STOPWORDS and len(token) > 3:
            result.append(token)
    return result

# print out the new extacted sub-topics from tne new doc based on the trained model
def pred_new_doc_extracted_topics(lda_mdl, bow_vector, wrd_n):
#     res_new_doc = lda_mdl[bow_vector]
#     for index, score in sorted(lda_mdl[bow_vector], key=lambda tup: -1*tup[1]):
#         print("Score: {}\t Topic: {}".format(score, lda_mdl.print_topic(index, wrd_n)))
    return lda_mdl[bow_vector]

refer_lda_models = pickle.load(open('models/trained_models/LDA_Subtopic_Models_ReferenceText.pckl',"rb"))

# count the number of sub-topics_traineds
main_topics_lst = []
cnt_extracted_subtopics_train = []
for lda_dict in refer_lda_models:
    key_val = list(lda_dict.keys())[0]
    
    main_topics_lst.append(key_val)
    
    cnt = 0
    for i in lda_dict[key_val][0].print_topics():
#         print(i)
        cnt += 1
    cnt_extracted_subtopics_train.append(cnt)

total_train_subtopic_df = pd.DataFrame({"main_topic" : main_topics_lst, 
              "total_train_subtopic_cnt" : cnt_extracted_subtopics_train })


total_train_subtopic_df

#### Subtopic/Concept/Fact/Relevance Measurement 

- count the # of subtopics lda modesl trained on reference text ==> n_trained_subtopics

- count the # of subtopics lda model applied on new text ==> n_test_subtopics_txt

- count the # of subtopics lda model applied on new sentences ==> n_test_subtopics_sentence

- overall subtopic coverage probability: 

- the probability of a subtopic on a new text ==> P_txt_subtopic_i

- the probability of a subtopic on a new sentences ==> P_sentences_subtopic_i



main_topics_lst = list(total_train_subtopic_df.main_topic)
# ['bird_tailorbird', 'insect_ant', 'bird_peacock', 'insect_beetle','bird_flamingo']

txt_subtopics_extracted_dict = {}

sentences_subtopics_extracted_dict = {}

train_lda_model_topics_dict = {}

for j in range(len(main_topics_lst)):
    
    print("#################################")
    print("main topics: {}".format(main_topics_lst[j]))
    print("------------------------------------")
    
    # read the pre-train lda model from a specific main topic
    lda_mdl = refer_lda_models[j][main_topics_lst[j]][0]
    
    # extract the trained model topics details and topics indexs
    lda_mdl_detail = lda_mdl.print_topics()
    
    # extract index and topics formula into lists
    trained_extra_subtopics_idx_lst = [i[0] for i in lda_mdl_detail]
    
    trained_extra_subtopics_lst = [i[1] for i in lda_mdl_detail]
    
    train_lda_model_topics_dict[main_topics_lst[j]] = []
    
    train_lda_model_topics_dict[main_topics_lst[j]].append(trained_extra_subtopics_idx_lst)
    
    train_lda_model_topics_dict[main_topics_lst[j]].append(trained_extra_subtopics_lst)
    
    # read the pre-train dictionary from a refernece text under specific main topic
    train_txt_dict_vec = refer_lda_models[j][main_topics_lst[j]][1]
    
    file_lst = list(df_rank[df_rank['main_topics'] == main_topics_lst[j]]['Title'].unique())

    for i in range(len(file_lst)):

        # read the whole txt transcript based on specific main topics
        new_transripts = list(df_rank[(df_rank['Title'] == file_lst[i])]['final_corrected_version_txt'].drop_duplicates())

        print(file_lst[i])

        # collect transcript 
        for new_transript in new_transripts:
            # apply the lda model on the new transcript
            bow_vector_txt = train_txt_dict_vec.doc2bow(preprocess(new_transript))

            pred_res_txt = pred_new_doc_extracted_topics(lda_mdl = lda_mdl, bow_vector = bow_vector_txt, wrd_n = 10)

            txt_subtopics_extracted_dict[file_lst[i]]= []

            tpc_idxs = []
            tpc_probas = []
            main_topics = []
            if pred_res_txt != []:
                for res in pred_res_txt:
                    tpc_idxs.append(res[0])
                    tpc_probas.append(res[1])
                    main_topics.append(main_topics_lst[j])

            txt_subtopics_extracted_dict[file_lst[i]].append(tpc_idxs)
            txt_subtopics_extracted_dict[file_lst[i]].append(tpc_probas)
            txt_subtopics_extracted_dict[file_lst[i]].append(main_topics)

        # read all the sentences based on specific main topics
        new_sentences_doc = list(df_rank[(df_rank['Title'] == file_lst[i])]['final_corrected_version_sentences_txt'])
        
        sentences_subtopics_extracted_dict[file_lst[i]] = []
        # apply the lda model on the new sentences
        for new_sentence in new_sentences_doc:

            bow_vector_sent = train_txt_dict_vec.doc2bow(preprocess(new_sentence))

            pred_res_sent = pred_new_doc_extracted_topics(lda_mdl = lda_mdl, bow_vector = bow_vector_sent, wrd_n = 10)

            one_sentence_dict = {}

            one_sentence_dict[new_sentence] = []

            tpc_idxs = []
            tpc_probas = []
            main_topics = []
            if pred_res_txt != []:
                for res in pred_res_txt:
                    tpc_idxs.append(res[0])
                    tpc_probas.append(res[1])
                    main_topics.append(main_topics_lst[j])
                    
            one_sentence_dict[new_sentence].append(tpc_idxs)
            one_sentence_dict[new_sentence].append(tpc_probas)
            one_sentence_dict[new_sentence].append(main_topics)
           
            sentences_subtopics_extracted_dict[file_lst[i]].append(one_sentence_dict)
#             print(sentences_subtopics_extracted_dict)




# based on the extracted subtopics and its stats from the LDA model on the Transcript text, formluate a dataframe

txt_subtopics_extracted_df = pd.DataFrame()

for key, ele in txt_subtopics_extracted_dict.items():
    
    txt_subtopics_extracted_df_temp = pd.DataFrame()
    
    txt_subtopics_extracted_df_temp["txt_subtopics_indx"] = ele[0]
    
    txt_subtopics_extracted_df_temp["txt_subtopics_proba"] = ele[1]
    
    txt_subtopics_extracted_df_temp["main_topic"] = ele[2]
    
    txt_subtopics_extracted_df_temp['Title'] = key
    
    txt_subtopics_extracted_df = pd.concat([txt_subtopics_extracted_df, txt_subtopics_extracted_df_temp], axis = 0)
    
txt_subtopics_extracted_df = txt_subtopics_extracted_df.reset_index(drop = True)

txt_groupby_df = txt_subtopics_extracted_df.groupby(["main_topic", 'Title'\
                                                          ]).agg({'txt_subtopics_indx': "count"\
                                                                 }).reset_index().rename(columns={'txt_subtopics_indx': 'txt_subtopics_count'})

# print(sent_groupby_df.columns)

txt_subtopics_extracted_df= pd.merge(txt_subtopics_extracted_df, txt_groupby_df, how = 'left', on = ['main_topic', 'Title'])

txt_subtopics_extracted_df= pd.merge(txt_subtopics_extracted_df, total_train_subtopic_df, how = 'left', on = ['main_topic'])

txt_subtopics_extracted_df['subtopic_cv_ratio'] = txt_subtopics_extracted_df['txt_subtopics_count']/ txt_subtopics_extracted_df['total_train_subtopic_cnt']

txt_subtopics_extracted_df.head()

txt_subtopics_extracted_df.to_csv("model_output_dataset/video_tarnscript_txt_subtopic_extraction.csv")

# based on the extracted subtopics and its stats from the LDA model on the Transcript Sentences, formluate a dataframe

# based on the extracted subtopics and its stats from the LDA model on the Transcript text, formluate a dataframe

sentence_subtopics_extracted_df = pd.DataFrame()

for key, eles in sentences_subtopics_extracted_dict.items():
    
    sentence_subtopics_extracted_df_temp = pd.DataFrame()
    
    for ele in eles:
    
        sentc = list(ele.keys())[0]

        sentence_subtopics_extracted_df_temp["sentences_subtopics_indx"] = ele[sentc][0]

        sentence_subtopics_extracted_df_temp["sentences_subtopics_proba"] = ele[sentc][1]

        sentence_subtopics_extracted_df_temp["main_topic"] = ele[sentc][2]

        sentence_subtopics_extracted_df_temp["sentences"] = sentc

        sentence_subtopics_extracted_df_temp['Title'] = key

        sentence_subtopics_extracted_df = pd.concat([sentence_subtopics_extracted_df, sentence_subtopics_extracted_df_temp], axis = 0)

sentence_subtopics_extracted_df= sentence_subtopics_extracted_df.reset_index(drop = True)

# print(sentence_subtopics_extracted_df.columns)

sent_groupby_df = sentence_subtopics_extracted_df.groupby(["main_topic", 'Title', 'sentences'\
                                                          ]).agg({"sentences_subtopics_indx": "count"\
                                                                 }).reset_index().rename(columns={'sentences_subtopics_indx': 'sentences_subtopics_count'})

# print(sent_groupby_df.columns)

sentence_subtopics_extracted_df= pd.merge(sentence_subtopics_extracted_df, sent_groupby_df, how = 'left', on = ['main_topic', 'Title', 'sentences'])

sentence_subtopics_extracted_df.head()


sentence_subtopics_extracted_df.to_csv("model_output_dataset/video_tarnscript_sentences_subtopic_extraction.csv")

trained_lda_subtopics_extracted_df = pd.DataFrame()

for key, ele in train_lda_model_topics_dict.items():
    
    trained_lda_subtopics_extracted_df_temp = pd.DataFrame()
    
    trained_lda_subtopics_extracted_df_temp["trained_subtopics_indx"] = ele[0]
    
    trained_lda_subtopics_extracted_df_temp["trained_subtopics_formula"] = ele[1]
    
    trained_lda_subtopics_extracted_df_temp["main_topic"] = key
    
    trained_lda_subtopics_extracted_df = pd.concat([trained_lda_subtopics_extracted_df, trained_lda_subtopics_extracted_df_temp], axis = 0)
    
trained_lda_subtopics_extracted_df= trained_lda_subtopics_extracted_df.reset_index(drop = True)

groupby_df = trained_lda_subtopics_extracted_df.groupby("main_topic").agg({"trained_subtopics_indx": "count"\
                                                                    }).reset_index().rename(columns={"trained_subtopics_indx": 'trained_subtopics_count'})

trained_lda_subtopics_extracted_df= pd.merge(trained_lda_subtopics_extracted_df, groupby_df, how = 'left', on = 'main_topic')

trained_lda_subtopics_extracted_df.head()
    

## join the sentences level new doc extrated probaility with the trained subtopics 

sentence_subtopics_extracted_eval_df = pd.merge(sentence_subtopics_extracted_df, trained_lda_subtopics_extracted_df,\
         left_on=['sentences_subtopics_indx','main_topic'], right_on=['trained_subtopics_indx','main_topic'], how='left')

sentence_subtopics_extracted_eval_df.head()


sentence_subtopics_extracted_eval_df.to_csv("model_output_dataset/sentence_subtopics_extracted_eval_df.csv")

## rank the txt subtopic extraction rate: new-doc-subtopic_n / main_topic_subtopic_n
sentence_subtopics_extracted_eval_df["new_sent_concept_coverage_pct"] = \
     np.round(sentence_subtopics_extracted_eval_df["sentences_subtopics_count"].astype(float) / sentence_subtopics_extracted_eval_df["trained_subtopics_count"], 4)


sentence_subtopics_extracted_eval_rank_df = sentence_subtopics_extracted_eval_df.sort_values(by = "new_sent_concept_coverage_pct")[['main_topic', 'sentences', 'Title', "new_sent_concept_coverage_pct"]].fillna(0)

## join the txt level new doc extrated probaility with the trained subtopics 

txt_subtopics_extracted_eval_df = pd.merge(txt_subtopics_extracted_df, trained_lda_subtopics_extracted_df,\
         left_on=['txt_subtopics_indx','main_topic'], right_on=['trained_subtopics_indx','main_topic'], how='left').fillna(0)

txt_subtopics_extracted_eval_df.head()


### Apply Given Main topics to select the top video

txt_subtopics_extracted_eval_df["new_txt_concept_coverage_pct"] = \
     np.round(txt_subtopics_extracted_eval_df["txt_subtopics_count"].astype(float) / txt_subtopics_extracted_eval_df["trained_subtopics_count"], 4)


# txt_subtopics_extracted_eval_rank = txt_subtopics_extracted_eval_df[["main_topic", "Title",  "subtopic_cv_ratio"]].drop_duplicates()

# txt_subtopics_extracted_eval_rank

txt_subtopics_extracted_eval_df


txt_subtopics_extracted_eval_df.to_csv("model_output_dataset/txt_subtopics_extracted_eval_df.csv")
