==> Pipelines/3.0_combine_validate_relevance_model_result_py38.ipynb <==
import sklearn
import pandas as pd
import numpy as np
import nltk
import matplotlib.pyplot as plt

df_rank = pd.read_csv("model_output_dataset/video_tarnscript_main_topic_classifier_rank.csv").drop('Unnamed: 0', axis = 1)
print(df_rank.columns)
print(df_rank.shape)
df_rank.head()

### Insights on the main topic classifier

under a main-topic, what is the top rank on the video files 

df_rank_classifier = df_rank[['Title', 'Video_ID', 'txt_main_topic_classified_proba', 'main_topics', 'coverage_quality_y']].drop_duplicates()
df_rank_classifier = df_rank_classifier[df_rank_classifier['coverage_quality_y'].notnull()]


df_rank_classifier

main_topics_Lst = list(df_rank_classifier.main_topics.unique())

for topic in main_topics_Lst:
    print()
    print('top 10 video under main topic: {} :'.format(topic))
    temp_df = df_rank_classifier[df_rank_classifier['main_topics'] == topic].sort_values(by = 'txt_main_topic_classified_proba', ascending = False)
    print()
    print(temp_df.head(10)['Video_ID'])

### Calcualte Subtopics Score KPIs 

import gensim
from gensim.utils import simple_preprocess
from gensim.parsing.preprocessing import STOPWORDS
from nltk.stem import WordNetLemmatizer, SnowballStemmer
from nltk.stem.porter import *
import numpy as np
# np.random.seed(2018)

import nltk
nltk.download('wordnet')

from Python_Functions.NLP_Model_Functions import stopwords_generator, lemmatize_stemming, TF_whole_text, TF_IDF_text_doc, topic_extract, \
                          lemmatize_preprocess, text_cleaning_sub, text_cleaning, stop_words

txt_subtopics_extracted_df = pd.read_csv("model_output_dataset/video_tarnscript_txt_subtopic_extraction.csv").drop('Unnamed: 0', axis = 1)
print(txt_subtopics_extracted_df.columns)
print(txt_subtopics_extracted_df.shape)
txt_subtopics_extracted_df.head()

sentence_subtopics_extracted_df= pd.read_csv("model_output_dataset/video_tarnscript_sentences_subtopic_extraction.csv").drop('Unnamed: 0', axis = 1)

print(sentence_subtopics_extracted_df.columns)
print(sentence_subtopics_extracted_df.shape)
sentence_subtopics_extracted_df.head()


txt_subtopics_extracted_eval_df = pd.read_csv("model_output_dataset/txt_subtopics_extracted_eval_df.csv").drop("Unnamed: 0", axis = 1)

print(txt_subtopics_extracted_eval_df.columns)
print(txt_subtopics_extracted_eval_df.shape)
txt_subtopics_extracted_eval_df.head()
                                             
                                             

## how to combine the result from Main topic classifier with Sub-topic Extraction


### Implement the probaility of a Sentences/Text covers a certain subtopic while this sentence/transcript txt also get covered by a main topic

P(A & B): the probaility a sentences/txt relevant to a subtopics and also get covered by the main topic

P(A & B) = P(A) *  P(B) 

txt_subtopics_extracted_eval_rank = txt_subtopics_extracted_eval_df[["main_topic", "Title", "subtopic_cv_ratio"]].drop_duplicates()

txt_subtopics_extracted_eval_rank


df_rank.columns

## text level relevance_score in sub_topic + main_topics

## keep the txt level classifier score only
txt_main_topics_classifier_rank = df_rank[['Title', 'Video_ID', 'final_corrected_version_txt',
       'main_topics_coded_x','txt_0_proba', 'txt_1_proba', 'txt_2_proba', 'txt_3_proba',
       'txt_4_proba', 'classified_topic_coded_txt', 'txt_main_topic_classified_proba',
       'if_classified_correctly_txt',  'max_proba_txt', 'txt_classified_vs_actual_gap']].drop_duplicates()

print(txt_main_topics_classifier_rank.shape)

text_level_relevance_score_df = pd.merge(txt_subtopics_extracted_eval_df, txt_main_topics_classifier_rank, \
                                          how = 'left', left_on = ['Title'], right_on = ['Title' ])

text_level_relevance_score_df = pd.merge(text_level_relevance_score_df, txt_subtopics_extracted_eval_rank, \
                                          how = 'left', left_on = ['Title',  'main_topic'], \
                                          right_on = ['Title' , 'main_topic'])


# calculate the text level releancy score
text_level_relevance_score_df['total_relevancy_score_proba_txt'] = text_level_relevance_score_df['txt_subtopics_proba'] * \
             text_level_relevance_score_df['txt_main_topic_classified_proba']


print(text_level_relevance_score_df.shape)



### Total Video File Txt level main topic relevance score: 
    
take the dot prudtc between the the vector of the extracted subtopic probaility with the vector of subtpics count ratio agins the total numbers of trained subtopics

- txt_main_topic_classified_proba: used to measure the probaility of a main topics covers the transcript text

- txt_subtopics_proba: used to measure a single subtopic relevance quality extracted from a transcript text; we main use this KPI to rank the video files subtopic relevance quality under a given transcript text


- total_txt_level_subtopics_relevance_score: is the score encaptuate and measure the total sub-topic extraction quality of a text transcript.

      total_txt_level_subtopics_relevance_score = dot(subtopic_proba, extracted_subtopic_count_ratio) 
      
- total_txt_level_relevance_score: is the score measure the quality by a total relavance score from both the main topics and their extracted subtopics; we main use this KPI to rank the video files overall relevance quality under a given main topics 

      weighted_total_txt_level_relevance_score = total_txt_level_subtopics_relevance_score * main_topic_classified_proba
      
      total_txt_level_relevance_score = txt_main_topic_classified_proba * main_topic_classified_proba

- total_txt_level_subtopic_level_relevance_score: measure the relavnce quality of a indvidual subtopic extracted from a transcript text and the proba of getiing covered by the main topic; we main use this KPI to rank the video files subtopic total relevance quality under a given main topics 

      txt_level_subtopics_level_relevance_score = subtopic_proba * main_topic_classified_proba

def total_subtopic_probab_score(DF, fl_nm,  proba_col, cnt_ratio, stences = None, txt_lvl = True):
    
    if txt_lvl == True:

        temp_DF = DF[DF['Title'] == fl_nm]
        
#         print(temp_DF)

        return np.round(np.dot(np.array(list(temp_DF[proba_col])), \
               np.array(list(temp_DF[cnt_ratio]))), 4)
    else: # sentences 
        
        temp_DF = DF[(DF['Title'] == fl_nm) & (DF['sentences'] == stences)]

        return np.round(np.dot(np.array(list(temp_DF[proba_col])), \
               np.array(list(temp_DF[cnt_ratio]))), 4)

txt_lvl_total_relevance_calcu = text_level_relevance_score_df[['Title', 'main_topic', 'txt_subtopics_indx', \
                                'txt_subtopics_proba', 'subtopic_cv_ratio_x', 'txt_main_topic_classified_proba']].drop_duplicates()

txt_lvl_total_relevance_calcu.head()

# calculate the total_txt_level_subtopics_relevance_score

file_names = list(txt_lvl_total_relevance_calcu['Title'].unique())
total_subtopic_relevance_scores = []

for file in file_names:
#     print(file)
    total_rel_score = total_subtopic_probab_score(\
                        DF = txt_lvl_total_relevance_calcu, fl_nm = file, \
                    proba_col = 'txt_subtopics_proba', cnt_ratio='subtopic_cv_ratio_x')
    total_subtopic_relevance_scores.append(total_rel_score)
#     print(total_rel_score)
    
txt_lvl_relevance_scores_df = pd.DataFrame({"Title" : file_names
            ,"total_txt_level_subtopics_relevance_score": total_subtopic_relevance_scores, 
             })
txt_lvl_relevance_scores_df.head()

# calculate total_txt_level_relevance_score
txt_lvl_total_relevance_calcu = pd.merge(txt_lvl_total_relevance_calcu, txt_lvl_relevance_scores_df, "left", \
         left_on = ['Title'], right_on = ['Title'])
txt_lvl_total_relevance_calcu['total_txt_level_relevance_score'] = txt_lvl_total_relevance_calcu['total_txt_level_subtopics_relevance_score'] * \
                                          txt_lvl_total_relevance_calcu['txt_main_topic_classified_proba'] 
txt_lvl_total_relevance_calcu.head()


##### given 5 main topics what are the top 10 video transcript txt in terms of total relavnce quality with total_txt_level_relevance_score

main_topics_lst = list(txt_lvl_total_relevance_calcu['main_topic'].unique())

for main_topic in main_topics_lst:
    print()
    print('--------------------------')
    print("Under a main_topic {},  we have our top 10 videos ranked by total_relevance_score: \n".format(main_topic))
    [print(i) for i in list(txt_lvl_total_relevance_calcu[txt_lvl_total_relevance_calcu['main_topic'] ==\
                                  main_topic][['Title', 'main_topic', \
                                'total_txt_level_relevance_score']].drop_duplicates().sort_values(\
                                    by = 'total_txt_level_relevance_score', ascending = False).head(10)['Title'])]
